ff
Update profile page:
1. Added a setter function to allow customization of the erasing speed in the animation.
2. Expanded the content displayed on the call-to-action package for better engagement.
3. Adjusted the opacity of the call-to-action section to improve visual appeal.
4. Removed redundant CSS code from the stylesheet to optimize performance and maintainability.
