+++ index.html
@@ -34,8 +34,8 @@
             <div class="container">
 
                 <div class="info">
-                    <h1 class="roboto-bold height-xl color-gray">Egbie Uku</h1>
-                    <span id="animated-text" class="text poppins-bold color-blue"></span>
+                    <h1 class="roboto-bold height-xl color-red">Egbie Uku</h1>
+                    <span id="animated-text" class="poppins-bold color-white"></span>
 
                 </div>
